test_anthropic_messages_litellm_router_non_streaming_with_logging
